requirements.txt
firebase-admin
python-dotenv
PyGithub
beautifulsoup4